Conflicts:
	src/main/webapp/WEB-INF/views/01/30/list.jsp

+++ src/main/java/kr/co/itcen/fa/controller/menu01/Menu03Controller.java
@@ -2,7 +2,12 @@ package kr.co.itcen.fa.controller.menu01;
 
 
 import java.text.ParseException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -13,6 +18,8 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 import kr.co.itcen.fa.dto.DataResult;
 import kr.co.itcen.fa.security.Auth;
 import kr.co.itcen.fa.security.AuthUser;
@@ -20,6 +27,8 @@ import kr.co.itcen.fa.service.menu01.Menu03Service;
 import kr.co.itcen.fa.service.menu17.Menu19Service;
 import kr.co.itcen.fa.service.menu17.Menu59Service;
 import kr.co.itcen.fa.vo.UserVo;
+import kr.co.itcen.fa.vo.menu01.ItemVo;
+import kr.co.itcen.fa.vo.menu01.MappingVo;
 import kr.co.itcen.fa.vo.menu01.VoucherVo;
 
 
@@ -74,7 +83,7 @@ public class Menu03Controller {
 		// 마감 여부 체크
 		
 		System.out.println("asdf: " + voucherVo.getRegDate());
-		
+		//String businessDateStr = menu03Service.businessDateStr();
 		if(menu19Service.checkClosingDate(userVo, voucherVo.getRegDate())) {
 			voucherVo.setOrderNo(1);
 			menu03Service.createVoucher(voucherVo, userVo);
@@ -82,31 +91,37 @@ public class Menu03Controller {
 		// 전표등록, 리스트
 		DataResult<VoucherVo> dataResult = menu03Service.selectAllVoucherCount(page);
 		model.addAttribute("dataResult", dataResult);
-		return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/list";
+		return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/read";
 	}
 	
 	// 전표 삭제 1팀
 	@RequestMapping(value = "/" + SUBMENU + "/delete", method=RequestMethod.POST)
-	public String delete(@ModelAttribute VoucherVo voucherVo, @AuthUser UserVo userVo) {
+	public String delete(@ModelAttribute VoucherVo voucherVo, @AuthUser UserVo userVo) throws ParseException {
 		if(!voucherVo.getInsertTeam().equals(userVo.getTeamName())) {
-			return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/list";
+			return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/read";
 		}
-		menu03Service.deleteVoucher(voucherVo);
 		
-		return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/list";
+		//String businessDateStr = menu03Service.businessDateStr();
+		if(menu19Service.checkClosingDate(userVo, voucherVo.getRegDate())) {
+			menu03Service.deleteVoucher(voucherVo);
+		}
+		
+		return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/read";
 	}
 	
 	// 전표 수정 1팀
 	@RequestMapping(value = "/" + SUBMENU + "/update", method=RequestMethod.POST)
-	public String update(@ModelAttribute VoucherVo voucherVo, @AuthUser UserVo userVo) {
+	public String update(@ModelAttribute VoucherVo voucherVo, @AuthUser UserVo userVo) throws ParseException {
 		voucherVo.setUpdateUserid(userVo.getId());
 		if(!voucherVo.getInsertTeam().equals(userVo.getTeamName())) {
-			return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/list";
+			return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/read";
+		}
+		System.out.println("orderNo : " + voucherVo.getOrderNo());
+		if(menu19Service.checkClosingDate(userVo, voucherVo.getRegDate())) {
+			menu03Service.updateVoucher(voucherVo);
 		}
 		
-		menu03Service.updateVoucher(voucherVo);
-		
-		return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/list";
+		return "redirect:/"+ MAINMENU + "/" + SUBMENU + "/read";
 	}
 	
 	// 거래처, 은행, 계좌, 카드 조회
@@ -117,5 +132,57 @@ public class Menu03Controller {
 		data.put("success", true);
 		return data;
 	}
+	
+	// 전표 추가
+	@ResponseBody
+	@RequestMapping(value = "/" + SUBMENU + "/save", method=RequestMethod.POST)
+	public Map<String, Object> save(HttpServletRequest request, @AuthUser UserVo userVo) {
+		Map<String, Object> resultMap = new HashMap<String, Object>();
+		String itemList = request.getParameter("itemList");
+		System.out.println(itemList);
+		
+		ObjectMapper mapper = new ObjectMapper();
+		try {
+			
+			VoucherVo[] voucherList = mapper.readValue(itemList, VoucherVo[].class);
+			
+//			System.out.println(voucherList);
+//			System.out.println(voucherList[0]);
+//			System.out.println(voucherList[0].getAccountName());
+//			System.out.println(voucherList[1]);
+//			System.out.println(voucherList[1].getAccountName());
+			
+			if(menu19Service.checkClosingDate(userVo, voucherList[0].getRegDate())) {
+				VoucherVo voucherVo = new VoucherVo();
+				voucherVo.setRegDate(voucherList[0].getRegDate());
+				
+				List<ItemVo> itemList2 = new ArrayList<ItemVo>();
+	 			
+				for(int i = 0; i < voucherList.length; i++) {
+					ItemVo itemVo = new ItemVo();
+					itemVo.setAmount(voucherList[i].getAmount());
+					itemVo.setAmountFlag(voucherList[i].getAmountFlag());
+					itemVo.setAccountNo(voucherList[i].getAccountNo());
+					
+					itemList2.add(itemVo);
+				}
+				
+				MappingVo mappingVo = new MappingVo();
+				
+				mappingVo.setVoucherUse(voucherList[0].getVoucherUse());
+				mappingVo.setCustomerNo(voucherList[0].getCustomerNo());
+				mappingVo.setDepositNo(voucherList[0].getDepositNo());
+				mappingVo.setManageNo(voucherList[0].getManageNo());
+				
+				menu03Service.createVoucher(voucherVo, itemList2, mappingVo, userVo);
+			}
+			
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		System.out.println("111");
+		return resultMap;
+	}
+	
 	
 }
